+++ src/service/task.ts
@@ -1,12 +1,15 @@
 /**
  * タスクサービス
  */
-import * as cinerino from '@cinerino/domain';
 import * as moment from 'moment';
 import * as mongoose from 'mongoose';
 import * as redis from 'redis';
 
+import { MongoRepository as TaskRepo } from '../repo/task';
+
 import * as factory from '@tokyotower/factory';
+
+import * as NotificationService from './notification';
 
 export interface IConnectionSettings {
     /**
@@ -19,7 +22,7 @@ export interface IConnectionSettings {
     redisClient: redis.RedisClient;
 }
 
-export type TaskOperation<T> = (repos: { task: cinerino.repository.Task }) => Promise<T>;
+export type TaskOperation<T> = (repos: { task: TaskRepo }) => Promise<T>;
 export type IOperation<T> = (settings: IConnectionSettings) => Promise<T>;
 
 export const ABORT_REPORT_SUBJECT = 'Task aborted !!!';
@@ -32,7 +35,7 @@ export function executeByName<T extends factory.taskName>(params: {
     name: T;
 }): IOperation<void> {
     return async (settings: IConnectionSettings) => {
-        const taskRepo = new cinerino.repository.Task(settings.connection);
+        const taskRepo = new TaskRepo(settings.connection);
 
         // 未実行のタスクを取得
         let task: factory.task.ITask<any> | null = null;
@@ -56,7 +59,7 @@ export function execute(task: factory.task.ITask<any>): IOperation<void> {
     const now = new Date();
 
     return async (settings: IConnectionSettings) => {
-        const taskRepo = new cinerino.repository.Task(settings.connection);
+        const taskRepo = new TaskRepo(settings.connection);
 
         try {
             // タスク名の関数が定義されていなければ、TypeErrorとなる
@@ -86,7 +89,7 @@ export function retry(params: {
     project?: factory.project.IProject;
     intervalInMinutes: number;
 }): TaskOperation<void> {
-    return async (repos: { task: cinerino.repository.Task }) => {
+    return async (repos: { task: TaskRepo }) => {
         await repos.task.retry(params);
     };
 }
@@ -101,7 +104,7 @@ export function abort(params: {
      */
     intervalInMinutes: number;
 }): TaskOperation<void> {
-    return async (repos: { task: cinerino.repository.Task }) => {
+    return async (repos: { task: TaskRepo }) => {
         const abortedTask = await repos.task.abortOne(params);
 
         // tslint:disable-next-line:no-single-line-block-comment
@@ -117,7 +120,7 @@ export function abort(params: {
             /* istanbul ignore next */
             '';
 
-        await cinerino.service.notification.report2developers(
+        await NotificationService.report2developers(
             ABORT_REPORT_SUBJECT,
             `project:${(params.project !== undefined) ? params.project.id : ''}
 id:${abortedTask.id}